build: Fix NetworkLayer
