+++ frontend/src/api/module/key.js
@@ -4,5 +4,5 @@ export const Key = {
   SESSION: 'session',
   LOCATION: 'location',
   STATUS: 'status',
-  TOKEN: 'token'
+  JOIN: 'session/join'
 }
